requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
pymysql==1.1.0
pydantic==2.6.0
python-multipart==0.0.6